% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advheidel.R
\name{plot.advHeidelDiag}
\alias{plot.advHeidelDiag}
\title{Plot Method for advHeidelDiag Class}
\usage{
\method{plot}{advHeidelDiag}(x, ...)
}
\arguments{
\item{x}{An object of class `advHeidelDiag` containing MCMC data and diagnostic results.}

\item{...}{Additional arguments (not used).}
}
\description{
This function generates diagnostic plots (trace, cumulative mean, half-width) for each chain.
}
