% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advheidel.R
\name{print.advHeidelDiag}
\alias{print.advHeidelDiag}
\title{Print Method for advHeidelDiag Class}
\usage{
\method{print}{advHeidelDiag}(x, ...)
}
\arguments{
\item{x}{An object of class `advHeidelDiag` containing diagnostic results.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints the Heidelberger-Welch diagnostic results in a table format,
including custom metrics and test type if available.
}
