% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advheidel.R
\name{summary.advHeidelDiag}
\alias{summary.advHeidelDiag}
\title{Summary Method for advHeidelDiag Class}
\usage{
\method{summary}{advHeidelDiag}(object, ...)
}
\arguments{
\item{object}{An object of class `advHeidelDiag`.}

\item{...}{Additional arguments (not used).}
}
\description{
This function provides a concise summary of the Heidelberger-Welch diagnostic results,
including the number of chains that passed or failed the stationarity and half-width tests.
}
